Updated build.gradle to use "implementation" rather than "compile"
Dave Syer pointed out that "compile" is outdated and that it should be "implementation" and Andy Wilkinson figured out how to make that work. (The odd thing is that I had tried it and couldn't get it to work. I must have had some other error, too.) I made the change Andy suggested, and it works now. Thanks, Dave and Andy. :)